+++ src/library/Base64CharacterEncodedByteSequence/definition.ts
